implement base file structure and styled() function
